Added files to repo
Using Kevin's example and the website redblobgames as reference, created a map that randomly generates certain terrain. Gave user options for size of map (not difficulty dependent), and has a method to ask user for difficulty. Gui class is if we want an actual image and the FastNoiseLite class creates noise to help generate a map. Nathan cleaned up the code, cleaned up the output and created a key for color for randomly generated item.